Move optimizer tests into separate file
These are getting quite large, so move them into a separate file.
